requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
factory_boy==3.3.1
Faker==30.0.0
freezegun==1.5.1
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg==3.2.2
psycopg-binary==3.2.2
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
